+++ R/rules-other.R
@@ -1,7 +1,6 @@
-#' @importFrom purrr reduce
 add_brackets_in_pipe <- function(pd) {
   is_pipe <- pd$token == "SPECIAL-PIPE"
-  reduce(which(is_pipe), add_brackets_in_pipe_one, .init = pd)
+  Reduce(add_brackets_in_pipe_one, which(is_pipe), init = pd)
 }
 
 add_brackets_in_pipe_one <- function(pd, pos) {
@@ -20,7 +19,7 @@ add_brackets_in_pipe_one <- function(pd, pos) {
       pd$child[[next_non_comment]],
       new_pd
     ) %>%
-      arrange(pos_id)
+      arrange_pos_id()
   }
   pd
 }
@@ -135,7 +134,7 @@ wrap_subexpr_in_curly <- function(pd,
     slice(-ind_to_be_wrapped) %>%
     bind_rows(new_expr_in_expr) %>%
     set_multi_line() %>%
-    arrange(pos_id)
+    arrange_pos_id()
 }
 
 #' Check if if, for or while loop expression require a braces.