+++ R/token-create.R
@@ -30,9 +30,11 @@ create_tokens <- function(tokens,
                           indention_ref_pos_ids = NA,
                           indents = 0,
                           terminal = TRUE,
-                          child = NULL) {
+                          child = NULL,
+                          stylerignore = FALSE) {
   len_text <- length(texts)
-  tibble(
+  new_tibble(
+    list(
     token = tokens,
     text = texts,
     short = substr(texts, 1, 5),
@@ -47,7 +49,10 @@ create_tokens <- function(tokens,
     multi_line = rep(FALSE, len_text),
     indention_ref_pos_id = indention_ref_pos_ids,
     indent = indents,
-    child = rep(list(child), len_text)
+    child = rep(list(child), len_text),
+    stylerignore = stylerignore
+  ),
+    nrow = len_text
   )
 }
 
@@ -89,7 +94,7 @@ find_start_pos_id <- function(pd, pos, by, direction, after, candidates = NULL) 
     ifelse(after, max(candidates), min(candidates)) + by * direction
   } else {
     find_start_pos_id(
-      pd$child[[pos]], if_else(after, nrow(pd$child[[pos]]), 1L),
+      pd$child[[pos]], ifelse(after, nrow(pd$child[[pos]]), 1L),
       by, direction, after, candidates
     )
   }